requirements.txt
Django==1.10.2
djangorestframework==3.4.7
pep8==1.7.0
